Extract parseValidatorsList
