+++ src/app/state/network/saga.ts
@@ -2,7 +2,7 @@ import * as oasis from '@oasisprotocol/client'
 import { PayloadAction } from '@reduxjs/toolkit'
 import { config } from 'config'
 import { call, put, select, takeEvery } from 'typed-redux-saga'
-import { AccountsApi, BlocksApi, Configuration, OperationsListApi } from 'vendors/explorer'
+import { getMonitorAPIs } from 'vendors/monitor'
 
 import { networkActions } from '.'
 import { selectSelectedNetwork } from './selectors'
@@ -36,16 +36,7 @@ function getBlocksPerEpoch(genesis: oasis.types.GenesisDocument): number {
 export function* getExplorerAPIs(network?: NetworkType) {
   const selectedNetwork = yield* select(selectSelectedNetwork)
   const url = config[selectedNetwork].explorer
-
-  const explorerConfig = new Configuration({
-    basePath: url,
-  })
-
-  const accounts = new AccountsApi(explorerConfig)
-  const blocks = new BlocksApi(explorerConfig)
-  const operations = new OperationsListApi(explorerConfig)
-
-  return { accounts, blocks, operations }
+  return getMonitorAPIs(url)
 }
 
 export function* selectNetwork({ payload: network }: PayloadAction<NetworkType>) {
